NillionBackend/base/profiles/serializers.py
from rest_framework import serializers
from base.models import Profile, Message, Chat

class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = Profile
        fields = '__all__'

    def validate_address(self, value):
        if Profile.objects.filter(address=value).exists():
            raise serializers.ValidationError("Address must be unique.")
        return value


class MessageSerializer(serializers.ModelSerializer):
    sender = ProfileSerializer(read_only=True)

    class Meta:
        model = Message
        fields = '__all__'

class ChatSerializer(serializers.ModelSerializer):
    participants = ProfileSerializer(many=True, read_only=True)
    messages = MessageSerializer(many=True, read_only=True)

    class Meta:
        model = Chat
        fields = '__all__'
