NillionBackend/requirements.txt
djongo==1.3.6
pymongo==3.12.3
django==4.1.9
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
django

pymongo[srv]
g4f

py-nillion-client
cosmpy
python-dotenv
nillion_python_helpers
django-cors-headers
djangorestframework